+++ src/components/Header.jsx
@@ -5,20 +5,26 @@ import { makeStyles } from '@material-ui/core/styles';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 
-// import IconButton from '@material-ui/core/IconButton';
-// import MenuIcon from '@material-ui/icons/Menu';
+import IconButton from '@material-ui/core/IconButton';
+import MenuIcon from '@material-ui/icons/Menu';
+import Hidden from '@material-ui/core/Hidden';
 
 // import { DifficultyButtons, DarkModeButton } from './Buttons';
 const DifficultyButtons = lazy(() => import('./buttons/DifficultyButtons'));
 const DarkModeButton = lazy(() => import('./buttons/DarkModeButton'));
-const LanguageSelector = lazy(() => import('../languages/Language'));
+const LanguageSelector = lazy(() => import('./buttons/LanguageSelector'));
 
 
 
 const useStyles = makeStyles((theme) => ({
     root: { flexGrow: 1, },
-    menuButton: { marginRight: theme.spacing(2), },
-    title: { flexGrow: 1, }
+    title: { flexGrow: 1, },
+    margin: {
+        marginRight: theme.spacing(2),
+        [theme.breakpoints.down('sm')]: {
+            marginRight: theme.spacing(0.5),
+        },
+    }
 }));
 
 const renderLoader = () => null;
@@ -29,18 +35,26 @@ export default function ButtonAppBar(props) {
         <div className={classes.root}>
             <AppBar position="static">
                 <Toolbar>
-                    {/* <IconButton edge="start" className={classes.menuButton} color="inherit" aria-label="menu">
+                    <IconButton edge="start" className={classes.menuButton} color="inherit" aria-label="menu">
                         <MenuIcon />
-                    </IconButton> */}
-                    <Typography variant="h6" className={classes.title}>Sudoku </Typography>
+                    </IconButton>
+                    <Hidden xsDown>
+                        <Typography variant="h6" className={classes.title}>Sudoku </Typography>
+                    </Hidden>
                     <Suspense fallback={renderLoader()}>
-                        <LanguageSelector />
+                        <div className={classes.margin}>
+                            <LanguageSelector />
+                        </div>
                     </Suspense>
                     <Suspense fallback={renderLoader()}>
-                        <DarkModeButton {...props} />
+                        <div className={classes.margin}>
+                            <DarkModeButton {...props} />
+                        </div>
                     </Suspense>
                     <Suspense fallback={renderLoader()}>
-                        <DifficultyButtons {...props} />
+                        <div className={classes.margin}>
+                            <DifficultyButtons {...props} />
+                        </div>
                     </Suspense>
                 </Toolbar>
             </AppBar>
